Updated openstack/openstack Project: openstack/cinder  7008b2a5b3a3cf590f71d6bb42102b35650bc9e4
Implementing the use of _L’x’/i18n markers

Placing the _Lx markers back into the code.  No other cleaner solution has
has been implemented. Patches will be submitted in a series of sub
directories and in a fashion that is manageable.
This is the fifth commit of this kind

Change-Id: I476750bf05ddee85dd5bf25c7bd00ee103d23115
Partial-Bug: #1384312