Updated openstack/openstack Project: openstack/cinder  3a691648c9c302e78e3b4430480a007c654fdd88
Mock isfile in test_ssh_missing_hosts_key_file

test_ssh_missing_hosts_key_file tests that when the ssh_hosts_key_file
is missing an InvalidInput exception is raised. It does this by
setting CONF.ssh_hosts_key_file to, what is assumed to be, the
location of a non-existant file. If the file, /tmp/blah, exists the
test fails. This test should not be dependent on what is on the
filesystem.

This commit adds a mock for os.path.isfile so that when
ssh_utils.SSHPool checks for existance of CONF.ssh_hosts_key_file it
always gets False regardless of the state of the filesystem.

Change-Id: I6bf797301a0829797d55d4442088f9ef944aec2c
Closes-bug: #1392628