Updated openstack/openstack Project: openstack-dev/devstack  e260a1050dfe24a607bc90c1cef5b9bc7f76d7c9
One last format and content cleanup for the index page

Update the form of the links to other documents in the tree;
the generated HTML links must remain in the original form.
The layout is changed from the previous tables to a more Sphinx-y
TOC-like list.

Also editorial changes to the index page and guide titles

Change-Id: I52df0cc542754c386fb3c99d9efdf2524c11bf48