Updated openstack/openstack Project: openstack/cinder  6dacf44f27966593b427f950fbdbb19e37bdf5bd
Add i18n _LX tags for relevant log levels in EQLX driver.

The decision has been made to follow the translation
convention used in other projects of marking all INFO
and above log strings with _LI, _LW, or _LE. This patch
adds these markers for all log strings in the EQLX
driver that are not being used in an exception or where
array output is being logged.

Change-Id: I978886e0a98601c0b3d70f7c97bccf6561ba325e