Updated openstack/openstack Project: openstack/nova  df2f71e325ff8869258585a81fa811b70a021487
Virt: update spawn signature to pass instance_type

This patch updates the spawn method to receive the instance type
as input. This will save database calls when spawning instances.

This is due to the fact that a number of the virt drivers are now
reading the flavor extra specs. There is no need to access the
database to get this information.

An example of where this will be used is
commit af4ec4ccb1ce5e635171c1d4b39f37fe0a5ccc34.

Change-Id: Ieebfcbc2478ba8d7036c1af30efbbf640aa7df73