Updated openstack/openstack Project: openstack/cinder  ede07344ca5a7ca23d703443eeffd7c9a3b4c182
allow image_id for imageRef in create volume API v2

The create volume request for the V2 API has a property named
"imageRef" with the ID of the image to be used when creating the
volume.
The name of this property is inconsistent as all multi word
properties use snake case and the property in the reponse is named
"image_id".

This adds support for aliasing the imageRef key with image_id when
creating a volume in v2 of the Cinder API in a backward-compatible
way.

Change-Id: I85a5c6734ba05a8a7ea9eff3f849d12fb3d167fb
Closes-Bug: 1375867