Updated openstack/openstack Project: openstack/requirements  e5d096d51a38d20c9e17dc0ec9e9469c4d28c9ba
Add Werkzeug>=0.7 to global-requirements.txt

Werkzeug>=0.7 is a direct dependency of Designate, and a transitive
dependency of the Flask WSGI framework used by Designate and
Ceilometer.

Werkzeug is a stable + actively maintained collection of WSGI
utilities, providing the basis for the Flask WSGI framework. The
codebase supports Python 2.6, 2.7 and 3.3, is licenced under
the BSD licnese, and is packaged on Ubuntu, Debian, Fedora.

Change-Id: I955dc03010c9988f0dc989217c5c084eede34fb5