Updated openstack/openstack Project: openstack-dev/devstack  99225de8cfe8bcb04f4259a3bbf53461d2dc1d12
make run_tests.sh run devstack unit tests

run_tests.sh was created as an entry point for bashate. We stopped
using it now that we know how to run bashate directly in the gate. 
We lost running an unrelated bashate test when that happened.

Devstack does have unit tests. We don't run them. We should. 
Especially when working on things like the LIBS_FROM_GIT which is
sufficiently tricky that we really need to add unit tests to ensure
that we don't break it.

Change-Id: Ife067855569e2eae4c085471d326e8086de37332