Updated openstack/openstack Project: openstack/cinder  ad75e5cdfd20d8a2cf54e7073554af84d737a0f1
Adding support for 'source-id' in 3PAR manage

Changed 3PAR's manage volume functionality to support the use of a
'source-id' id-type.  By default the name of the volume to manage can
be used but if an admin wants to use the volume ID instead they can
set the optional --id-type argument to 'source-id' now.

Closes-Bug: 1357075
Change-Id: I503beefc594f03627d2273f48c80326cf80fbdef