Updated openstack/openstack Project: openstack/cinder  c0fb19e96ee4c22212fd72ec44933ce282dbe014
Fix typo in SolidFire xDBVersionMismatch label

The retryable errors list in the SolidFire driver includes
an xDBVersionMismatch error that can be emitted from the
device.  During some recent work however the spelling of
the error was incorrect, so it's never deteceted/retried
as it should be.

This patch fixes the typo from 'xDBVersionMisMatch' to
xDBVersionMismatch'.

Change-Id: I5b4bf908c6a9146c4554e5749cd12dbc46c6e7c1
Closes-Bug: #1393916