Updated openstack/openstack Project: openstack/nova  90dee8d431380d3cc2eb87c558585be1e3880b7a
Fixes differencing VHDX images issue on Hyper-V

When using differencing vhdx images, one will be created pointing
to the base image when a VM is spawned. This image will be created
having the according flavor size.

In this case, due to a Win32 API issue, the image is not generated
properly.

Simply creating the root image with no size specified and resizing
it afterwards fixes the issue.

Change-Id: I7bb46a63d4c2a06ed6f63c839461cbbc1abffa65
Closes-Bug: #1370531