Updated openstack/openstack Project: openstack/neutron  5ddeabeabff9983f5baa463819ef92c99d0d732c
BSN: Set inconsistency record on delete failure

Set a bad value in the consistency DB on a delete
failure so a sync is triggered the next time the backend
servers become available.

This fixes the issue where the backend servers are unavailable
when an object is deleted from ML2 but not deleted from the
backend.

Closes-Bug: #1377350
Change-Id: I24713a03af4d3499645241f900de1f572689c235