Updated openstack/openstack Project: openstack/oslo.db  027bdcb303e9c2327aa86ae1f504dd960a43801c
Add pbr to installation requirements

Add pbr to the list of installation requirements so that it is installed
via pip before this library is installed, instead of with easy_install.
This avoids issues like Bug #1384919, and ensures that projects that use
this library as a dependency are properly installed.

Related-Bug: #1384919

Change-Id: I6c155370dbd01fe4748d5137bdf288e8d3e1a67e