Updated openstack/openstack Project: openstack/oslo-incubator  a09320a702ae3d7a1abf4c7151a3528b247ddc73
threadgroup: don't log GreenletExit

This exception is always raised when a green thread is stopped. This
exception does not indicate any problem in a service, so it should not
be logged not to scare operators and developers that waste their time
debugging those exceptions in their deployment and unit test logs.

Change-Id: I03dd3484430166839219af4d4881ea404526ce72
Related-Bug: #1382573