Updated openstack/openstack Project: openstack/horizon  1b0a8883902be169311ff2366bda506058be3ffc
Add unit conversion for metering views

Until now, the charts displayed by the metering views (Admin -> Resource
Usage) were using units as defined by the Ceilometer meters. This means
that the values in the charts were sometimes very hard to read for
humans (typically we would see a large number in nanoseconds, bytes,
etc).

This patch introduces support for normalization of these values. The
statistics returned by the Ceilometer API are analyzed to determine
the appropriate unit for the series. The whole series is then
converted to this unit, so that the values in the series are more
easily parsed by humans.

Implements blueprint: ceilometer-api-statistics-decorators-and-convertors
Change-Id: Ia183ca5270ebd1dcbd1fe6ac20ed4dfe4d6ebcc4