Updated openstack/openstack Project: openstack/tempest  9b149413b79709cffc04b17698ebba801b39e391
Fix inheritance for BaseTestCase

This commit fixes the inheritance around the BaseTestCase definition.
Previously there were a couple of things wrong with it, first the
mixin order was incorrect, secondly the use of ResourcedTestCase
wasn't actually needed since testresources isn't actually used
anywhere. The other part of the fix was the hack for python2.6 to
also add unittest2 to BaseTestCase is not needed now, because
testtools now uses unittest2.

As a part of this fix the base CLI test class has to be reworked to
change how tempest-lib is used. By fixing the inheritance on
BaseTestCase a conflict occurs with the class test used from
tempest-lib. The recommended usage pattern for tempest-lib will have
to be rethought. But, in the mean time this commit just adds the code
which would be inherited from the mixin in tempest-lib directly. It
should be removed when tempest-lib is updated to accommodate this use
case.

Change-Id: Ia5dda59f7989f37997276cbd19bb9fdc7b7d2624