Updated openstack/openstack Project: openstack/heat  ee1683e1ba22ba593768d1a75d69abc6057c10ce
Make ResourceGroup updates update nested stacks

When you update a ResourceGroup which contains a provider resource
any update only happens when the properties input to the provider
change.  Instead, always trigger an update, so we can respond to
changes to the provider template definitions which may occur.

For AutoScalingGroup, we may want something more sophisticated,
where instead of unconditionally recurising on update, the option
to instead use rolling updates to do batched replacemnt of nested
stacks is offered.  But for ResourceGroup (which is a simple
container resource), this behavior seems most appropriate.

Change-Id: I6e0e50c821a27a0a9fb16dbf186558da3632935d
Closes-Bug: #1389499