Updated openstack/openstack Project: openstack/cinder  1d77d79f81cf5e6a808453432f6e510ad0ea5a25
Implementing the use of _L’x’/i18n markers

Placing the _Lx markers back into the code.  No other cleaner solution has
has been implemented. Patches will be submitted in a series of sub
directories and in a fashion that is manageable.
This is the sixth commit of this kind

Partial-Bug: #1384312

Change-Id: I42b4e168deec9930571c1869fe1a181d4aad1112