Updated openstack/openstack Project: openstack/taskflow  53895c12ee76984dfce8731bd1ee7d0ddc213820
Mark 'task_notifier' as renamed to 'atom_notifier'

Deprecate the usage of an engines 'task_notifier' in
favor of the more appropriately named 'atom_notifier' and
mark the 'task_notifier' property as subject to removal
in a future version.

This makes the usage of this notifier more clear since it
is not only used for task notification but also for retry
notification (which is why naming it atom notifier is more
suited to its actual usage).

Change-Id: Idfa31a6665e43dcc9360f0aed40e27817d9e6737