Updated openstack/openstack Project: openstack/glance  3f0b479fed34a2bc312e58def73b00b205fab13b
Handle some exceptions of image_create v2 api

Currently, in image_create v2 api source we can see
ReservedProperty, ReadonlyProperty and TypeError caused in
glance/domain/__init__.py are not handled, which will cause
500 when these exceptions happens.The internal error should
be interpreted to the appropriate exception to the user.
This patch catches them and changes them to http exception.

Change-Id: I0c82a142b3357c8b8faab7dc0905967983cb8f7c