Updated openstack/openstack Project: openstack/nova  ecbe5357e171518819e6113ed126cf5dfc3897bc
VMware: remove ESX support for getting resource pool

The ESX driver is no longer supported. We should only return the
resource pool for the cluster.

In order to test this the patch had to update the test_vmops to
ensure that a cluster was used.

The patch also drops the method get_mo_id_from_instance which is
no longer used.

Change-Id: I9d99a5be58ef5a2a7d41dfd488ac24f38d2c79d3