Updated openstack/openstack Project: openstack/tempest  9be1c2350555b9ac3a80737ce3c142289c7a0260
Add scenario test for swift-proxy ceilometer middleware

This test creates a swift container, puts something in it
and then checks for samples stored by ceilometer which were
created by notifications from the middleware. test.call_until_true
is used to loop across multiple requests for samples in order
to be flexible in the face of latency on the notifications bus.

Change-Id: I07316db1f51d0562919ed2a50f002011944bfdb1