Updated openstack/openstack Project: openstack/trove  b2aa4f993e4559fee0410bd7b69753e4ab8c0e2c
Deleting failed replication backup can hide error

If there is a problem with getting the replication snapshot, the backup
is deleted. If the delete fails (if, for example, the backup is in the
wrong state) then the original exception is lost (and more importantly,
not logged).

Handling is now in place to trap exceptions while deleting a 'bad'
replication backup. All exceptions are logged and re-thrown as a
TroveError.

The duplicated code around the backup delete has also been collapsed
into one (nested) exception.

Change-Id: I858df7fd7eeb79dd1d559c4dd4b2851f7d0e0d98
Closes-Bug: #1386901