Updated openstack/openstack Project: openstack/swift  308c6a5e73ad8f9c1c669d7406b7f55bf46e1af0
mock out os.listdir to return a list

os.listdir returns a list of items.  The test case had been
written to return a single item which, though not really changing
the result of the test, was not the best approach.

This patch updates the test case to return a list instead of a single
item.

Change-Id: I793e0636440c0de0ca339c6592adec3e8b4ee1b4