Updated openstack/openstack Project: openstack/neutron  6e1be2d32ed7256de6d284fcea7f942fd4348ac1
Prevent an iteration through ports on IPv6 slaac

A recent change[1] allowed subnets to be deleted even if they
had active IPv6 SLAAC allocations on them. The updated check
was inefficient because it would check every port in the subnet
even if the subnet was a SLAAC subnet. This patch just shortcuts
out that check.

1. I281f5a1553248e09174dc49d0a42aef4b5c44bee

Change-Id: I2c35495b3642c644e4758f28ccddcc076139ec3b