Updated openstack/openstack Project: openstack/cinder  3ec99175818d01eaad09c7947c862593846f3b51
context.elevated() should use copy.deepcopy()

Currently context.elevated is just doing a copy.copy(self).
This needs to be changed to use copy.deepcopy so that the
list reference is not shared between objects which leaves
the possibility of an admin role leak.

This fix changes context.elevated use copy.deepcopy.

Change-Id: I349c53ccbe9e02ad2a3e84ae897424db9785a170
Closes-bug: 1386932