Updated openstack/openstack Project: openstack/sahara  27a08dc1274fc16a28a819fbeeb45b75d61550d5
Format volumes filesystems in parallel

Formatting in ext4 can be a long operation (several minutes for large
disks). With multiple volumes per instance, this significantly increases
cluster set-up time.

Since formatting and mounting can be done in parallel (bottleneck is
disk IO, not CPU), spawn a thread for each volume.

This change applies to the Heat engine only, because the direct engine
also attach devices to instances and this must be done sequentially.

Change-Id: Ie0d47c0906da7c123383646842003675096a2fc7
Closes-Bug: #1394593