Updated openstack/openstack Project: openstack/nova  fcba6c9a7aefed887f701bbc9f220d31918644e9
Revert "Truncate encoded instance message to 255"

This reverts commit bd9c32289ff8cf1bc67c1395ad921c703aa4c489.

UTF-8 character bytes size varies from 1 to 6. If truncating a long
bytes string to 255, the last character may be cut in the middle,
so that UnicodeDecodeError will occur when converting it back to
unicode. Need to revert this code change and make a new patch.

Related-Bug: 1389102

Change-Id: Ie439cffe11d657e77397c9070a3923a61f68a678