Updated openstack/openstack Project: openstack/cinder  fcd643ba636455d76f27eed711fab4fb3ece24bb
Implementing the use of _L’x’/i18n markers

Placing the _Lx markers back into the code. No other cleaner solution has
has been implemented. Patches will be submitted in a series of sub
directories and in a fashion that is manageable.
This is the fourth commit of this kind

Change-Id: Ibbef7f06a391e9e6efca082d45caecdf60a9e811
Partial-Bug: #1384312