Updated openstack/openstack Project: openstack/cinder  e17ac584f57c869dd3b90704fb57f5495b3639a3
PureISCSIDriver needs to disconnect hosts before deleting volumes.

Some error conditions can cause a situation where the volume is
“connected” to a Purity host, but the volume failed to attach on the
initiator side. If we then try to delete this volume it will cause
errors and leave orphaned volumes behind on the Flash Array.

The solution is to check for any connections and then remove them prior
to making a call to delete the volume.

Closes-Bug: 1388260
Change-Id: I8f5b3b28900d79228fd91f2ad53535224f263657