Updated openstack/openstack Project: openstack/keystone  ad7f02b8701bf319913256128d019cfe50c0afb4
Extends hacking check for logging to verify i18n hints

The original check would make sure that debug logging didn't use the
translation function. The check will now make sure that none of the
translation helpers are used for debug logging and that the correct
helper is used for logging statements.

bp more-code-style-automation

Change-Id: I536de5f21fe82fa3426040b8a4cfa5b81bdd5a8b