Updated openstack/openstack Project: openstack/nova  e022c65be07a1522a3a39eab1299f7cd6636d938
Limit tcp/udp port to be empty string in json-schema

tcp/udp port can be string and integer. Currently the json-schema accept
string, but it didn't limit the length of string, so there can be pass
emtpy string as tcp/udp port. This patch add string length limit to it.

Change-Id: I6c338a2417ae5f8817b24823b3d745dc1e0fbd58