Updated openstack/openstack Project: openstack/neutron  d42055cf214355d57026405045488ff7faaf68b7
Implement migration of legacy routers to distributed

This patch addresses the necessary changes required
for the router migration or conversion from legacy
to distributed.

The _update_router_db() method was modified to add
logic to support the SNAT functionality of DVR by
calling create_snat_intf_ports_if_not_exists().  A
call to _unbind_router() was also called to force
the scheduling of the router with 'distributed'
enabled.

Right now we will be supporting migration from legacy
to distributed only.

Closes-bug: #1348309
Partially-implements: blueprint neutron-ovs-dvr
Co-Authored-By: Michael Smith <michael.smith6@hp.com>
Change-Id: I6f252271bb5b52ce57184ad7b0bf1ce280b965fb